[Refactor] StarrySkyView 를 재사용해 온보딩 뷰에도 적용했습니다.
