+++ star/star/Source/Presentation/Onboarding/OnboardingCell.swift
@@ -47,8 +47,7 @@ final class OnboardingCell: UICollectionViewCell {
     // MARK: - Setup UI
     
     private func setupUI() {
-        guard let backgroundImage = UIImage(named: "backgroundImage") else { return }
-        backgroundColor = UIColor(patternImage: backgroundImage)
+        backgroundColor = .clear
         
         contentView.addSubviews(
             mockupImageView,
